Jsonhpp will parse everything as either int64 or uint64. We will stick with their convention and drop int in favor of int64_t in our internal settings format.

+++ box/Software/Triggering/lib/Settings.hpp
@@ -10,7 +10,7 @@
 
 namespace opq {
 
-    typedef boost::variant<uint64_t, float, int, std::string, bool> OPQSetting;
+    typedef boost::variant<uint64_t, float, int64_t, std::string, bool> OPQSetting;
 /**
  * @brief The Settings class synchronized singleton key value pair database.
  */
@@ -87,7 +87,7 @@ namespace opq {
         * @param key string identifier.
         * @return value.
         */
-        int getInt(std::string key);
+        int getInt64(std::string key);
         /**
          * @brief Get a string value for a key.
          * @param key string identifier.
